Cleanup: backport simplification from no-draggable/ to 18/ 19/ 20/ 21/
The change in coordinate systems means that the calculation for
the measure line left/center/right is no longer awkward so there's
much less justification for having it in the JS file. I only realized
this while writing no-draggable/ so now I'm backporting it to these
sections of the page.